tdf#148683 sw: fix crash on deleting text with redlining enabled, but hidden
Regression from commit 85d31a5d2cc1e425b82aa45e93de507e1e61d5f8
(sw_redlinehide: make layout based Show/Hide mode the default,
2018-11-30), deleting some text in the middle of the paragraph with the
bugdoc results in an assertion failure in the SwDrawTextInfo ctor.

Normally this doesn't happen on text deletion as we already have a
mechanism for truncating no longer needed lines in
SwTextFrame::Format_(), though that only checks the length of the
current line.

Fix the problem by extending this mechanism to also check the remaining
lines: if change tracking is on and the total of all line lengths is
more than the string length of the text frame, then time to
re-calculate.

It seems this is not needed in practice when redlining is disabled,
though we might want to enable this for the non-redline case as well in
the future.

Change-Id: Iede03e11daceb3c2b614a301e21560b075a60c01
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133620
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins