don't lay out entire strings for Writer layout tests
The SalLayoutGlyphsCache optimization of laying out an entire string
if it looks like a caller will call it repeatedly for parts of the string
conflicts with the SAL_ABORT_ON_NON_APPLICATION_FONT_USE checks
causing PrintFontManager::Substitute() abort if font fallback happens.
Simply disable the optimization, the output should be the same.

Change-Id: I0cce6a0c2a1a2ce93b95df54d18fedbdc779f34f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133789
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>