lay out entire strings in SalLayoutGlyphsCache more often (tdf#148911)
The document has parts of paragraphs e.g. underlined, which changes
the vcl::Font part of the hash key. So if a paragraph starts that way
but the rest is not underlined, the optimization of laying out
the entire string would be missed. So do the optimization if there
are subsequent calls for adjacent parts of the string (or starting
at the same index and different length).

Change-Id: I3d4a78b0eae42bd3085e96024340e6ed8daa9ad1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133790
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>