tdf#148818 Fix missing text in boxes with empty first paragraph
Text was lost from the text boxes with empty first paragraph. The
problem was happening when there was 2 paragraphs in a text box, in
which the first paragraph was completely empty. The regression was
introduced in b732004a50e005649ffdcbd50dbf1a33c4913581 because of the
bad handling of the paragraphs, and only looking at the size of the
text of the first paragraph.

This is fixed by looking at mpOutliner->GetEditEngine().GetTextLen(i)
for all i=1..k where k=min(2,mpOutliner->GetParagraphCount()).

I have negated the condition to provide a better explanation of what
the condition should be.

Change-Id: Id5d8787df5111c734760afdd98a6fbe832047f32
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133527
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>