add tests of query iterator's BinarySearch()
The BinarySeach() function is rather strange/dumb. If it doesn't
find the value, it tries to use the last previous value (i.e. it
additionally does the fixing up that VLOOKUP, MATCH, etc.) do.
At the some time, if there's a range of equal values, it doesn't
find the last one (and so callers need to fix that up). Write
some tests for that before I start touching the algorithm.

Change-Id: Ife2388acad691cce7ffaf5490fa74b5a3d453926
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134097
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>