ofz#47289 attempt to allocate negative len
a problem since...

commit 5a11d3888704b8e8a9137ff5634fefa369fea5a2
Date:   Sat Apr 2 15:49:32 2022 +0300

    tdf#103954: Z compressed graphic formats support for EMF/WMF

which added an unconditional seek to 0 of the stream which isn't
desirable when done on the original stream which is deliberately
seeked to the start of the graphic data, seeking to 0 just skips
back to the start of the enclosing file, e.g. a word document stream.

Presumably this should just happen in the case of the replacement
decompressed stream.

Change-Id: Ifb67fb265c70d5728a74dc3499f275eb5d69ddbb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133991
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>