sw XHTML export: output table / table row background format using CSS
The HTML export uses HTML elements / attributes to describe the
background color of tables / table rows. CSS markup is used when
exporting styles, not direct formatting.

This behavior makes sense for the HTML mode, but XHTML wants to CSS
markup whenever possible, and usage of CSS markup for the reqif mode is
not even optional.

Fix the problem by switching to CSS markup for table and table row
backgrounds in the XHTML case -- this is OK for reqif when describing
tables (the reqif spec only forbids detailed CSS markup for spans, not
tables or table rows).

This amends the behavior of commit
1f9915ce16e9c2ef4b2bcf53fbf2569ba512d3f1 (sw reqif-xhtml export: limit
values of the style attribute, 2018-04-11), which avoided invalid table
row background markup by losing it on export.

Change-Id: Ia0858986c3e8a3ea41adf8a24119442fe5068ee3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134119
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins