tdf#122057: read tiff with deflate compression (code 32946 only)
tiff format is more a multiformat container with different combination of parameters,
above all compression and photometric interpretation.

For this one, the only pb was deflate compression wasn't dealt with.
So just use existing LO wrapper for deflate lib in package part.

Remark: for the moment only decompression 32946 (legacy deflate) is managed not Adobe deflate (code 8).
I expected the same behaviour but some tests with tdf#131199 showed it didn't work.

Change-Id: I2c8d244fa89a2378bfe3b87d3d3262810c9952be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134485
Tested-by: Jenkins
Reviewed-by: Julien Nabet <serval2412@yahoo.fr>