sw content controls, picture: replace placeholder image on click
Previous content control types (rich text, checkbox, dropdown) had the
cursor somewhere in the content control and click selected the content
control in the placeholder case.

The picture case is somewhat different: the content control has a single
as-char image, and by the time SwWrtShell::GotoContentControl() is
invoked, the picture frame is already selected.

All we have to do here is to trigger the file picker to select a
replacement and then insert that image to replace the already selected
frame's bitmap.

Change-Id: Id3ad94ed8f7e13b0ebc2f57327bb8c0b1e374a86
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134511
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>