qt6 configure: Consider more paths to find Qt 6 moc
    commit dd52e7deae47a0fcb206a8187a95e60fba1484c8
    Date:   Wed Sep 29 11:09:51 2021 +0200

        qt6: Add a qt6 VCL plugin

mentioned:

> 1) At least in my self-compiled Qt versions,
> 'moc' (the meta-object compiler) is located
> in the 'libexec' subdirectory in 'QT6DIR', while the
> Qt 5 equivalent is located in the "bin" subdirectory
> of 'QT5DIR', so the configure.ac check uses the former.

Search for moc in the "libexec" subdirectory of
all paths in `lib_dir` before trying `$PATH`,
so e.g. the one provided as `/usr/lib/qt6/libexec/moc`
by package "qt6-base-dev-tools" in Debian testing is found
before a Qt 5 version in `$PATH`.

With this change in place, building on Debian testing
using the system Qt 6 works when building with autogen option
`--enable-qt6` but no `QT6DIR` set and `/usr/bin/moc` points
to the Qt 5 version of moc.

Change-Id: I0471c73315224ffab018b350aff8d797a234a905
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134814
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>