do not use glyph subsets if LTR/RTL does not match (tdf#149264)
Writer apparently sometimes asks to lay out RTL text as LTR,
for whatever reason, and Harfbuzz in that case does not give
consistent results usable for creating subsets of glyphs (or maybe
HB_GLYPH_FLAG_UNSAFE_TO_BREAK is not reliable in that case).
I don't quite understand what the problem here is, but avoid
it by checking whether the text is LTR or RTL and simply do
not try to create a subset from the glyphs if the text does
not match the way it was asked to be laid out. I don't even
know if this is correct handling or just a workaround for something.

Change-Id: Ic77db04f529c9ca2c194893a2127e85953accf32
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/134950
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>