tdf#149170 framework,sfx2: don't add ReadOnly=false
There is a non-obvious difference in error handling when opening the
file read/write for type-detection fails in
utl::MediaDescriptor::impl_openStreamWithURL() - if ReadOnly=false is
given, it will abort the file loading, whereas without any ReadOnly it
will fall-back to a read-only opening of the file for type detection.

At this point a StillReadWriteInteraction is used, so the user is never
notified of the failure.

(regression from commit 50affcfe2a1b2d19da7b09507fcb517221fe33d1)

Change-Id: I8c06fe23cc0bc0767df83f680a1a59e3700bbbbb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135045
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>