vcl: rename ImplAnimView to AnimationRenderer
An ImplAnimView is actually just a way of rendering an animation to an
OutputDevice (RenderContext). Each instance has a unique ID (called
mnExraData, which is quite misleading). ImplAnimView doesn't really
mean much so hence the rename to make this a bit more clear.

In the process, also rename local variables pView to pRenderer in
Animation functions.

Change-Id: Ib70f41052c5266acd7f2460001ce5be75b74d5a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/76402
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>