tdf#149431 Fix the sidebar's width
When click the wider sidebar and then go back to the narrow one in Writer,
the narrow one becomes the same width as wider one, it's not suitable
and doesn't look good. Fix the width to smallest width unless you drag
the splitter. Take the maximum width from the smallest width and
mnWidthOnSplitterButtonDown.

Change-Id: I60aaab84593d59ac28b96c2d3c3bd62a4bbddbbe
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135220
Tested-by: Jenkins
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>