sw content controls: update the cursor at SwContentControlDlg end
Once a dropdown content control is inserted, the properties dialog can
modify list items. The desktop case updates these list items when
SfxViewFrame::Enable() calls SwView::ShowCursor(), but for some reason
the same implicit update doesn't happen in the LOK case.

This means adding new list items requires leaving the content control &
entering again, then the dropdown shows the new list items, which is
confusing. And also the update (when it works) is implicit, so it can
break without us noticing.

Fix the problem by explicitly updating the cursor before ending the
properties dialog: this will scroll to the cursor position, but that's
not a problem since the properties dialog is always editing the content
control at the current position anyway.

Show / hide is used because that's an easy way to call
SwCursorShell::UpdateCursor(), which is private to us.

Change-Id: I43aca24c2961ef6c4cc14f895edd6f3f22980148
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135483
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins