avoid uninitialized data when handling WindowState
I get valgrind reports when I start LO with kf5 backend and
a maximized window, since in that case position and size are not
set (since the maximized state makes them irrelevant). The semantics
here seem to conflict, the caller of GetWindowState() sets flags
to say what data it wants, and SalFrame::GetWindowState() sets flags
to say what data it provides, but as can be seen here those aren't
necessarily the same. So actually use only those flags that are
in both sets.

Change-Id: I4559baab90d6f465382560549d52431b28119b3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135563
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>