sw XHTML export: avoid writing background of table cells in ReqIF mode
ReqIF mostly forbids using CSS styling on elements:
IgnorePropertyForReqIF() only allows 2 CSS keys by default.

This restriction was relaxed in commit
8e0b4ce9c538c51e079c5087262582787baf97f3 (sw XHTML export: output table
/ table row background format using CSS, 2022-05-10), to allow
background for tables and table rows, using CSS markup. An unwanted side
effect of this is background for table cells, which is still considered
invalid. To make this nontrivial to fix, Css1Background::Table is used
to track formatting for all of tables, rows and cells.

Fix the problem by extending Css1Background with a TableRow and
TableCell and then audit all uses of Css1Background::Table to explicitly
say if they mean table, row or cell. This keeps table and row
backgrounds, but fixes the unwanted cell background.

Also document the 3 functions doing the export of table / row / cell
background export to improve readability.

Change-Id: I03301b1fd25593cbe83489dbf140e80138d4a0de
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/135570
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins